refactor(oppijanumero): move OppijanumeroServiceError to an own file.
