+++ server/src/main/kotlin/fi/oph/kitu/oppijanumero/YleistunnisteHae.kt
@@ -1,7 +1,6 @@
 package fi.oph.kitu.oppijanumero
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import java.util.Date
 
 data class YleistunnisteHaeRequest(
     @JsonProperty("etunimet")
@@ -19,15 +18,4 @@ data class YleistunnisteHaeResponse(
     val oid: String,
     @JsonProperty("oppijanumero")
     val oppijanumero: String?,
-)
-
-data class OppijanumeroServiceError(
-    @JsonProperty("timestamp")
-    val timestamp: Date,
-    @JsonProperty("status")
-    val status: Int,
-    @JsonProperty("error")
-    val error: String,
-    @JsonProperty("path")
-    val path: String,
 )
